requirements.txt
streamlit>=1.35.0
ultralytics>=8.2.0
torch>=2.3.0
torchvision>=0.18.0
sentence-transformers>=3.0.0
transformers>=4.40.0
Pillow>=10.3.0
numpy>=1.26.0
opencv-python-headless>=4.9.0